requirements.txt
discord.py==2.3.2
virtualenv==20.25.1
python-dotenv==1.0.1
